Restructure folders

+++ include/lstd/common/cpp/source_location.h
@@ -1,9 +1,7 @@
 #pragma once
 
-#include "../numeric.h"
-
-// Use this to get the location where a function was called without using
-// macros.
+// Use this to get the location where a function
+// was called without using macros.
 
 #if defined LSTD_NO_CRT
 
@@ -12,7 +10,7 @@
 struct source_location {
   const char *File = "Unknown";
   const char *Function = "Unknown";
-  s64 Line = 0;
+  int Line = 0;
 
   constexpr source_location() {}
 
@@ -20,7 +18,7 @@ struct source_location {
   static consteval source_location
   current(const char *file = __builtin_FILE(),
           const char *func = __builtin_FUNCTION(),
-          s64 line = __builtin_LINE()) {
+          int line = __builtin_LINE()) {
     source_location loc;
     loc.File = file;
     loc.Function = func;
